Arrays

Arrays

CUDAdrv provides a primitive but useful array type to manage GPU data organized in an plain, dense fashion.

CUDAdrv.CuArray — Type.

CuArray{T}(dims)
CuArray{T,N}(dims)

Construct an uninitialized N-dimensional dense CUDA array with element type T, where N is determined from the length or number of dims. dims may be a tuple or a series of integer arguments corresponding to the lengths in each dimension. If the rank N is supplied explicitly as in Array{T,N}(dims), then it must match the length or number of dims.

Type aliases CuVector and CuMatrix are available for respectively 1 and 2-dimensional data.

source

Base.copy! — Method.

copy!{T}(dst::CuArray{T}, src::Array{T})

Copy an array from a host array src to a device array dst in place. Both arrays should have an equal length.

source

Base.copy! — Method.

copy!{T}(dst::Array{T}, src::CuArray{T})

Copy an array from a device array src to a host array dst in place. Both arrays should have an equal length.

source

Base.copy! — Method.

copy!{T}(dst::CuArray{T}, src::CuArray{T})

Copy an array from a device array src to a device array dst in place. Both arrays should have an equal length.

source

CUDAdrv.CuArray — Method.

CuArray{T}(src::Array{T})

Transfer a host array src to device, returning a CuArray.

source

Core.Array — Method.

Array{T}(g::CuArray{T})

Transfer a device array src to host, returning an Array.

source